Move /models directory into /api
