+++ api/services/users_service.go
@@ -6,7 +6,7 @@ import (
 
 	auth "github.com/decentorganization/topaz/api/core/authentication"
 	"github.com/decentorganization/topaz/api/core/database"
-	"github.com/decentorganization/topaz/models"
+	"github.com/decentorganization/topaz/api/models"
 )
 
 func NewUser(newUser *models.User) (int, []byte) {
